+++ package.json
@@ -30,6 +30,7 @@
     "emoji-regex": "^8.0.0",
     "entities": "^2.2.0",
     "escape-string-regexp": "^1.0.5",
+    "express": "^4.18.3",
     "external-editor": "^3.1.0",
     "figures": "^3.2.0",
     "follow-redirects": "^1.15.3",
@@ -63,8 +64,8 @@
     "safe-buffer": "^5.2.1",
     "safer-buffer": "^2.1.2",
     "signal-exit": "^3.0.7",
-    "string-width": "^4.2.3",
     "string_decoder": "^1.3.0",
+    "string-width": "^4.2.3",
     "strip-ansi": "^6.0.1",
     "supports-color": "^5.5.0",
     "through": "^2.3.8",
